Fix mbed_enter_sleep function.
The previous version of mbed_enter_sleep_function was only working if the
softdevice was enabled. The root cause of this bug is that WFE was called
after sd_softdevice_is_enabled.
A call to sd_softdevice_is_enabled trigger an SVC call and register an
event in the event register. When WFE is called, if an event is
registered, the cpu will not go to sleep.
This means that in the previous version of mbed_enter_sleep, if the
softdevice was not enabled, the CPU never goes to sleep.

The new startegy here is to clear the event register then, if there is
interrupts pending, just return from this function. If no interrupts are
pending, just call WFE.