Fix crash when interrupt is more than 8.5 minutes in the future
Basically the last condition in the if can be false if you set an interrupt too far in the future (more than one 24-bit timer loop), and if that happens then the IRQ is never cleared and the handler loops.

Fixes bug #53. I think. There are no tests so it's hard to be sure.